Add "url" file for entries
It's not possible to determine original entry URL for cache item because
file name is just sha256.

This adds extra 'url' file (similar to mime-type) that will store source
URL. For 'preload' case it's stored when image is saved. For 'proxy'
case it's also stored during GET if it was not stored before (some sort
of 'migration')